finish service and dao

+++ src/main/java/com/wisn/dao/ResourceDao.java
@@ -5,7 +5,7 @@ import com.wisn.entity.Resource;
 public interface ResourceDao {
     Resource queryResourceByResourceid(long resourceid);
 
-    int insertResource(Resource resource);
+    long insertResource(Resource resource);
 
     int deleteResourceByResourceId(long Resourceid);
 }
